+++ packages/fcl-ethereum-provider/src/cadence.ts
@@ -1,6 +1,20 @@
 import {getContractAddress} from "./util/eth"
 import {ContractType} from "./constants"
 
+export const getCOAScript = (chainId: number) => `
+import EVM from ${getContractAddress(ContractType.EVM, chainId)}
+
+access(all)
+fun main(address: Address): String? {
+    if let coa = getAuthAccount(address)
+        .storage
+        .borrow<&EVM.CadenceOwnedAccount>(from: /storage/evm) {
+        return coa.address().toString()
+    }
+    return nil
+}
+`
+
 export const createCOATx = (chainId: number) => `
 import EVM from ${getContractAddress(ContractType.EVM, chainId)}
 
@@ -14,6 +28,35 @@ transaction() {
 
         let cap = signer.capabilities.storage.issue<&EVM.CadenceOwnedAccount>(storagePath)
         signer.capabilities.publish(cap, at: publicPath)
+    }
+}
+`
+
+export const sendTransactionTx = (chainId: number) => `
+import EVM from ${getContractAddress(ContractType.EVM, chainId)}
+
+/// Executes the calldata from the signer's COA
+transaction(evmContractAddressHex: String, calldata: String, gasLimit: UInt64, value: UInt256) {
+    
+    let evmAddress: EVM.EVMAddress
+    let coa: auth(EVM.Call) &EVM.CadenceOwnedAccount
+
+    prepare(signer: auth(BorrowValue) &Account) {
+        self.evmAddress = EVM.addressFromString(evmContractAddressHex)
+
+        self.coa = signer.storage.borrow<auth(EVM.Call) &EVM.CadenceOwnedAccount>(from: /storage/evm)
+            ?? panic("Could not borrow COA from provided gateway address")
+    }
+
+    execute {
+        let valueBalance = EVM.Balance(attoflow: value)
+        let callResult = self.coa.call(
+            to: self.evmAddress,
+            data: calldata.decodeHex(),
+            gasLimit: gasLimit,
+            value: valueBalance
+        )
+        assert(callResult.status == EVM.Status.successful, message: "Call failed")
     }
 }
 `
